put my filters on the dinoKennel javascript file so that now my data just handlels my data.
