+++ javascripts/main.js
@@ -1,12 +1,12 @@
-import { getDinos } from './components/dinoData.js';
+import { dinoData } from './components/dinoData.js';
 import { buildDinoCards } from './components/dinoKennel.js';
 
 
 
 
 const init = () => {
-    buildDinoCards();
-    getDinos();
+    buildDinoCards(dinoData);
+  
 };
 
 init();
